feat: Add updateWorkflowTriggered flag for form submission
The code changes introduce a new boolean flag called `updateWorkflowTriggered` in the `_InfoCollectionPageState` class. This flag is used to track whether the form submission workflow has been triggered. It is set to `false` initially and updated to `true` when the submit button is pressed. If any of the required fields are empty, an error message is displayed using `Get.snackbar`. After the form submission is complete, the flag is set back to `false`. This enhancement improves the user experience by providing visual feedback during the form submission process.