feat: Update ItemLabel widget to include customizable label
This commit updates the ItemLabel widget to include a customizable label that can be displayed after the count of items. The new `label` parameter allows developers to provide a specific label to be shown, or it defaults to "Items" if no label is provided. This enhancement provides more flexibility in designing the appearance of the ItemLabel widget.