feat: Update MealBoxCard widget with customizable image position and scale
This commit adds new optional parameters to the MealBoxCard widget, allowing developers to customize the position and scale of the image within the card. The `topPos` parameter sets the top position of the image, defaulting to 0.1. The `leftPos` parameter sets the left position of the image, defaulting to 0.1. The `angle` parameter sets the angle of the image, defaulting to 0. The `scale` parameter sets the scale of the image, defaulting to 1.1. These enhancements provide more flexibility in designing the layout and appearance of the MealBoxCard widget.