feat: Add new SVG assets for arrowHead and dinner plate
This commit adds two new SVG assets, arrowHead.svg and dinner plate.svg, to the lib/assets/images and lib/assets/icons directories respectively. These assets are SVG files that can be used in the application for various purposes. The arrowHead.svg asset is a vector graphic of an arrowhead shape, while the dinner plate.svg asset is a vector graphic of a dinner plate. These assets enhance the visual elements and provide more options for designing the user interface.